Fix algorithm bugs, add usability features, CI/CD improvements
Correctness:
- V-trace: add dones parameter for episode boundary handling
- DreamerV3: fix gradient detach, use trajectory log_probs
- MAPPO: guard critic input for n_agents > 1
- OnlineDPO: fix CPU tensor loss accumulation
- DPO/GRPO/OnlineDPO: add gradient clipping (max_grad_norm=1.0)
- DPO: fix duplicate loss kwarg in callback

Usability:
- Add predict() to SAC, TD3, DQN
- Add ProgressBarCallback and TimingCallback
- Add CLI entry point (python -m rlox train/eval)
- torch.compile: compile individual policy methods
- MultiGPU: wrap off-policy networks, move targets
- Custom env: SAC/TD3/DQN accept str | gymnasium.Env
- SACTrainer: forward callbacks to algorithm
- GymVecEnv: float32 observations

Testing:
- Add 18 new tests for untested algorithms

CI/CD:
- Add Windows + Intel Mac wheels
- Build for Python 3.10-3.13
- Add ruff Python linting
- Remove broken python-publish.yml

+++ python/rlox/gym_vec_env.py
@@ -72,11 +72,11 @@ class GymVecEnv:
             for i in range(self._n_envs):
                 if mask[i] and infos["final_obs"][i] is not None:
                     terminal_obs[i] = np.asarray(
-                        infos["final_obs"][i], dtype=np.float64
+                        infos["final_obs"][i], dtype=np.float32
                     )
 
         return {
-            "obs": np.asarray(obs, dtype=np.float64),
+            "obs": np.asarray(obs, dtype=np.float32),
             "rewards": np.asarray(rewards, dtype=np.float64),
             "terminated": np.asarray(terminated, dtype=np.uint8),
             "truncated": np.asarray(truncated, dtype=np.uint8),
@@ -88,13 +88,13 @@ class GymVecEnv:
 
         Returns
         -------
-        np.ndarray of shape ``(n_envs, obs_dim)`` with dtype float64.
+        np.ndarray of shape ``(n_envs, obs_dim)`` with dtype float32.
         """
         kwargs: dict[str, Any] = {}
         if seed is not None:
             kwargs["seed"] = seed
         obs, _info = self._env.reset(**kwargs)
-        return np.asarray(obs, dtype=np.float64)
+        return np.asarray(obs, dtype=np.float32)
 
     def num_envs(self) -> int:
         """Return the number of parallel sub-environments."""
